feat(core): no more spek
