+++ core/src/integrationTest/kotlin/com/malinskiy/marathon/analytics/metrics/remote/influx/InfluxDbProviderIntegrationTest.kt
@@ -1,0 +1,66 @@
+package com.malinskiy.marathon.analytics.metrics.remote.influx
+
+import com.malinskiy.marathon.analytics.external.influx.InfluxDbProvider
+import com.malinskiy.marathon.analytics.external.influx.InfluxMetricsProvider
+import com.malinskiy.marathon.execution.AnalyticsConfiguration
+import com.malinskiy.marathon.generateTest
+import org.amshove.kluent.shouldEqualTo
+import org.influxdb.InfluxDB
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import java.time.Instant
+import java.time.temporal.ChronoUnit
+
+class InfluxDbProviderIntegrationTest {
+    private val database = "marathonTest"
+    private val rpName = "rpMarathon"
+
+    val container = KInfluxDBContainer().withAuthEnabled(false)
+
+    var thirdDbInstance: InfluxDB? = null
+
+    @BeforeEach
+    fun `start influx`() {
+        container.start()
+    }
+
+    @AfterEach
+    fun `stop influx`() {
+        thirdDbInstance?.close()
+        container.stop()
+    }
+
+    @Test
+    fun `multiple creations should still have the same configured retention policy`() {
+        val test = generateTest()
+        val provider = InfluxDbProvider(
+            AnalyticsConfiguration.InfluxDbConfiguration(
+                url = container.url,
+                dbName = database,
+                password = "",
+                user = "root",
+                retentionPolicyConfiguration = AnalyticsConfiguration.InfluxDbConfiguration.RetentionPolicyConfiguration.default
+            )
+        )
+        val firstDbInstance = provider.createDb()
+        firstDbInstance.close()
+
+
+        val secondDbInstance = provider.createDb()
+        prepareData(secondDbInstance, test, database, rpName)
+        secondDbInstance.close()
+
+        thirdDbInstance = provider.createDb()
+
+        val metricsProvider =
+            InfluxMetricsProvider(InfluxDBDataSource(thirdDbInstance!!, database, rpName))
+
+        val result = metricsProvider.executionTime(
+            test,
+            50.0,
+            Instant.now().minus(2, ChronoUnit.DAYS)
+        )
+        result shouldEqualTo 5000.0
+    }
+}
