+++ core/src/test/kotlin/com/malinskiy/marathon/time/SystemTimerTest.kt
@@ -4,35 +4,34 @@ import com.nhaarman.mockitokotlin2.reset
 import com.nhaarman.mockitokotlin2.whenever
 import org.amshove.kluent.mock
 import org.amshove.kluent.shouldEqualTo
-import org.jetbrains.spek.api.Spek
-import org.jetbrains.spek.api.dsl.describe
-import org.jetbrains.spek.api.dsl.it
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
 import java.time.Clock
 
-class SystemTimerSpek : Spek(
-    {
+class SystemTimerTest {
+    private val clock = mock<Clock>()
+    private val timer = SystemTimer(clock)
 
-        val clock = mock<Clock>()
-        val timer = SystemTimer(clock)
-        beforeEachTest {
-            reset(clock)
+    @BeforeEach
+    fun `reset mocks`() {
+        reset(clock)
+    }
+
+    @Test
+    fun `should call passed clock to get currentTimeMillis`() {
+        whenever(clock.millis()).thenReturn(100)
+
+        timer.currentTimeMillis()
+            .shouldEqualTo(100)
+    }
+
+    @Test
+    fun `should call passed clock to measure`() {
+        var counter = 0L
+        whenever(clock.millis()).thenAnswer {
+            counter++ * 1000L
         }
 
-        describe("SystemTimer") {
-            it("should call passed clock to get currentTimeMillis") {
-                whenever(clock.millis()).thenReturn(100)
-
-                timer.currentTimeMillis()
-                    .shouldEqualTo(100)
-            }
-
-            it("should call passed clock to measure") {
-                var counter = 0L
-                whenever(clock.millis()).thenAnswer {
-                    counter++ * 1000L
-                }
-
-                timer.measure { }.shouldEqualTo(1000L)
-            }
-        }
-    })
+        timer.measure { }.shouldEqualTo(1000L)
+    }
+}
